Revamp backend into a StatsDatabase
