+++ door-open-service/node-backend/src/app.js
@@ -1,47 +1,50 @@
+import fs from "fs";
 import histogram from "bars";
 import moment from "moment";
 import _ from "underscore";
-import fs from "fs";
 
-import backend from "./backend";
 import DoorSocketServer from "./DoorSocketServer";
 import DoorSlackBot from "./DoorSlackBot";
 import startHttpApi from "./httpApi";
+import StatsDatabase from "./StatsDatabase";
 
 const config = JSON.parse(fs.readFileSync("config.json"));
 
-const sockServer = new DoorSocketServer({
-  port: config.backend.port
-});
-
+const sockServer = new DoorSocketServer({ port: config.backend.port });
 const bot = new DoorSlackBot(config.slack);
 
 bot.onMessageLike(/garage/i, user => {
   if (!user) {
     bot.postMessage("Unrecognized user.");
-  } if (sockServer.available("garage")) {
-    const res = backend.openGarage(user, config.httpApi.secret, sockServer, config.httpApi.secret);
-    if (res) bot.postMessage(`Opening the garage as requested by ${user.name} (${user.email})...`);
-  } else {
+
+  } if (!sockServer.available("garage")) {
     bot.postMessage("The remote garage opening service is not operational at the moment. " +
         "Consider dispatching a drone to pick up a human.");
+
+  } else {
+    sockServer.broadcast("garage", "2500");
+    StatsDatabase.registerGarageOpen(user);
+    bot.postMessage(`Opening the garage as requested by ${user.name} (${user.email})...`);
   }
 });
 
 bot.onMessageLike(/open/i, user => {
   if (!user) {
     bot.postMessage("Unrecognized user.");
-  } else if (sockServer.available("door")) {
-    const res = backend.openDoor(user, config.httpApi.secret, sockServer, config.httpApi.secret);
-    if (res) bot.postMessage(`Opening the door as requested by ${user.name} (${user.email})...`);
-  } else {
+
+  } if (!sockServer.available("door")) {
     bot.postMessage("The remote door opening service is not operational at the moment. " +
         "Consider dispatching a drone to pick up a human.");
+
+  } else {
+    sockServer.broadcast("door", "2500");
+    StatsDatabase.registerDoorOpen(user);
+    bot.postMessage(`Opening the door as requested by ${user.name} (${user.email})...`);
   }
 });
 
 bot.onMessageLike(/stats/i, () => {
-  backend.getStats().then(stats => {
+  StatsDatabase.getStats().then(stats => {
     if (!stats) {
       throw new Error("There are no available stats for the remote door opening service");
     }
@@ -61,4 +64,4 @@ bot.onMessageLike(/stats/i, () => {
   });
 });
 
-startHttpApi(config, bot, sockServer, backend);
+startHttpApi(config, bot, sockServer);
